// src/app/models/Vehicle.ts
export interface Vehicle{
    stav?: string;
    vykonMotora?: number;
    pocetRychlosti?: number;
    pocetAirbagov?: number;
    karoseria?: string;   
    pocetDveri?: number;
    pocetKilometrov?: number;
    rokVyroby?: number;
    model?: string;
    cena?: number;
    objemMotora?: number;
    typPaliva?: string;


}